+++ package.json
@@ -25,12 +25,10 @@
     "@types/glob": "^7.1.1",
     "@typescript-eslint/eslint-plugin": "^2.25.0",
     "@typescript-eslint/parser": "^2.25.0",
+    "eslint": "^6.8.0",
+    "jest": "^25.1.0",
     "standard": "^14.3.3",
     "typescript": "^3.8.3"
-  },
-  "dependenciess": {
-    "rimraf": "^3.0.0",
-    "tslib": "^1.10.0"
   },
   "devDependenciess": {
     "@types/jest": "^24.0.19",
