+++ src/secure-rm.ts
@@ -1,111 +1,48 @@
-import crypto from 'crypto'
-import rimraf from 'rimraf'
-import glob from 'glob'// eslint-disable-line no-unused-vars
-import { standards, validIDs, Standard } from './standards' // eslint-disable-line no-unused-vars
-import { tree } from './events'
+import fs from 'fs-extra'
+import glob from 'glob'
 
-// to get the correct tree for each call
-function getUUID () {
-  return crypto.randomBytes(60).toString('base64').replace(/\//g, '0').replace(/\+/g, 'a')
-}
-
-interface Opts {
-  customStandard?: Standard
-  maxBusyTries?: number
-  emfileWait?: number
-  disableGlob?: boolean
+interface Options {
+  standard?: {
+    unlink?: typeof fs.unlink
+    rmdir?: typeof fs.rmdir
+  },
+  maxBusyTries?: number,
+  emfileWait?: number,
+  disableGlob?: boolean,
   glob?: glob.IOptions | false
 }
 
-export interface Options extends Opts {
-  standard?: keyof typeof standards
-}
-
-export interface ParsedOptions extends Opts {
-  standard: keyof typeof standards
-}
-
-export type Callback = (err: NodeJS.ErrnoException | null, tree: string[]) => void
-
-// Main function when secure-rm is called
-export function remove(path: string, options?: Options): Promise<string[]>
-export function remove(path: string, callback: Callback): void
-export function remove(path: string, options: Options, callback: Callback): void
-
-export function remove (path: string, options?: Options | Callback, callback?: Callback) {
-  const uuid = getUUID()
-  // Parse if callback is provided
-  if (callback === undefined && typeof options === 'function') {
-    callback = options
-    options = { standard: 'secure' }
-  }
-  // Define standard if none is provided
-  if (options === undefined) options = { standard: 'secure' }
-  if ((options as Options).standard === undefined) (options as Options).standard = 'secure'
-
-  if (callback) removeCallback(path, options as ParsedOptions, uuid, (err: NodeJS.ErrnoException | null, goodTree: string[]) => callback!(err, goodTree))
-  else return removePromise(path, options as ParsedOptions, uuid)
-}
+type Callback = (err: NodeJS.ErrnoException | null, res: any) => void
+type ThenArg<T> = T extends PromiseLike<infer U> ? U : T
+type ReturnPromise = ReturnType<typeof remove_>
+type ReturnCallback = ThenArg<ReturnType<typeof remove_>>
 
 const defaultGlobOpts = {
+  dot: true,
   nosort: true,
   silent: true
 }
 
-// Callback version
-function removeCallback (path: string, options: ParsedOptions, uuid: string, callback: Callback): void {
-  if (options.customStandard) {
-    rimraf(path, {
-      unlink: options.customStandard.unlinkStandard(uuid),
-      rmdir: options.customStandard.rmdirStandard(uuid),
-      maxBusyTries: options.maxBusyTries || 3,
-      emfileWait: options.emfileWait || 1000,
-      disableGlob: options.glob === false ? true : options.disableGlob || false,
-      glob: options.glob || defaultGlobOpts
-    }, (err: NodeJS.ErrnoException) => callback(err, tree[uuid]))
-  } else if (validIDs.includes(options.standard)) {
-    rimraf(path, {
-      unlink: standards[options.standard].unlinkStandard(uuid),
-      rmdir: standards[options.standard].rmdirStandard(uuid),
-      maxBusyTries: options.maxBusyTries || 3,
-      emfileWait: options.emfileWait || 1000,
-      disableGlob: options.glob === false ? true : options.disableGlob || false,
-      glob: options.glob || defaultGlobOpts
-    }, (err: NodeJS.ErrnoException) => callback(err, tree[uuid]))
-  } else {
-    callback(new Error(`'${options.standard}' is not a valid ID. \nList of valid IDs: ${validIDs}`), ['none'])
+export function remove (path: string, options?: Options): ReturnPromise
+export function remove (path: string, callback: Callback): ReturnCallback
+export function remove (path: string, options: Options, callback: Callback): ReturnCallback
+
+export function remove (path: string, options?: Options | Callback, callback?: Callback): ReturnPromise | ReturnCallback {
+  if (callback === undefined && typeof options === 'function') {
+    callback = options
+    options = undefined
   }
+  if (callback) {
+    remove_(path, options as Options)
+      .then(result => callback!(null, result))
+      .catch(err => callback!(err, null))
+  } else return remove_(path, options as Options)
 }
 
-// Promise version
-function removePromise (path: string, options: ParsedOptions, uuid: string): Promise<string[]> {
-  return new Promise((resolve, reject) => {
-    if (options.customStandard) {
-      rimraf(path, {
-        unlink: options.customStandard.unlinkStandard(uuid),
-        rmdir: options.customStandard.rmdirStandard(uuid),
-        maxBusyTries: options.maxBusyTries || 3,
-        emfileWait: options.emfileWait || 1000,
-        disableGlob: options.glob === false ? true : options.disableGlob || false,
-        glob: options.glob || defaultGlobOpts
-      }, (err: NodeJS.ErrnoException) => {
-        if (err) reject(err)
-        else resolve(tree[uuid])
-      })
-    } else if (validIDs.includes(options.standard)) {
-      rimraf(path, {
-        unlink: standards[options.standard].unlinkStandard(uuid),
-        rmdir: standards[options.standard].rmdirStandard(uuid),
-        maxBusyTries: options.maxBusyTries || 3,
-        emfileWait: options.emfileWait || 1000,
-        disableGlob: options.glob === false ? true : options.disableGlob || false,
-        glob: options.glob || defaultGlobOpts
-      }, (err: NodeJS.ErrnoException) => {
-        if (err) reject(err)
-        else resolve(tree[uuid])
-      })
-    } else {
-      reject(new Error(`'${options.standard}' is not a valid ID. \nList of valid IDs: ${validIDs}`))
-    }
+async function remove_ (path: string, options?: Options) {
+  await fs.remove(path, {
+    ...options,
+    ...options?.standard,
+    glob: options?.glob || defaultGlobOpts
   })
 }
