+++ test/junk/passes.test.js
@@ -1,0 +1,2 @@
+test.todo('Multiples passes')
+test.todo('Then function')
