+++ test/test.js
@@ -1,0 +1,4 @@
+import srm from '../dist'
+
+srm('./junk')
+  .then(() => console.log('success'))
